Move DefaultIngressFieldsFormik to common as it is used in Day2 context in OSD clusterDetails too
Update imports accordingly
