+++ src/components/clusters/wizards/rosa/NetworkScreen/NetworkScreen.jsx
@@ -22,9 +22,9 @@ import { getSelectedAvailabilityZones } from '~/common/vpcHelpers';
 import { constants } from '~/components/clusters/common/CreateOSDFormConstants';
 import { SubnetSelectField } from '~/components/clusters/common/SubnetSelectField';
 import { canConfigureDayOneManagedIngress } from '~/components/clusters/wizards/common/constants';
+import { DefaultIngressFieldsFormik } from '~/components/clusters/wizards/common/NetworkingSection/DefaultIngressFieldsFormik';
 import { useFormState } from '~/components/clusters/wizards/hooks';
 import { FieldId } from '~/components/clusters/wizards/rosa/constants';
-import { DefaultIngressFieldsFormik } from '~/components/clusters/wizards/rosa/NetworkScreen/DefaultIngressFieldsFormik';
 import { CheckboxDescription } from '~/components/common/CheckboxDescription';
 import ExternalLink from '~/components/common/ExternalLink';
 import { RadioButtons, ReduxCheckbox } from '~/components/common/ReduxFormComponents_deprecated';
